Use exception message copy to prevent dangling ref
Storing the const references result in dangling references as the
passed-in const references' lives do not span beyond the caller's stack
frame. The objects they reference are destroyed when the callers' stack
frames unwind. This is a property of C++ as const reference to
temporaries only persist as long as it is still within the calling
function; storing the const references further into a member won't
persist the temporary objects (the temporary life span is only
temporarily granted in the local scope and is not transitive).