Stop storing copies of actual and expected value
This allows expect(..) and matcher(..) of values that cannot be copied
(copy constructor deleted).

For temporary values, the C++ const reference spec allows a short
extended life cycle so it should be no problem to CCSpec.

```
The second context is when a reference is bound to a temporary. The
temporary to which the reference is bound or the temporary that is the
complete object to a subobject of which the temporary is bound persists
for the lifetime of the reference except as specified below. A temporary
bound to a reference member in a constructor’s ctor-initializer (§12.6.2
[class.base.init]) persists until the constructor exits. A temporary
bound to a reference parameter in a function call (§5.2.2 [expr.call])
persists until the completion of the full expression containing the
call.
```