Correct before and after hook example in README
The example code has a bug: the raw Duck pointer is not correctly
initialized in the before each block as the before each lambda takes a
copy of the pointer, therfore not altering the saved address of the
Duck pointer. Examples below would simply receive a dangling pointer.

Taking the reference of the Duck pointer would not work either, since
the raw Duck pointer will be destroyed when the eclosing example group
 lambda finishes, so the nested examples who reference it would receive
dangling pointers.

The correct way would be to use a shared pointer which has an extended
life cycle and each example takes a copy of the shared pointer.